requirements.txt
rich
click
pymodbus
python-snap7
BAC0
cpppo